[nasa/nos3#378] Split device into a new register file set;
